requirements.txt
# Core libraries
pyspark==3.3.2
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1
tqdm==4.65.0

# YouTube and video processing
pytubefix==4.0.0
yt-dlp==2023.7.6
google-api-python-client==2.88.0
fake_useragent==1.1.3

# Audio processing
faster-whisper==0.6.0
ffmpeg-python==0.2.0

# NLP and ML
transformers==4.30.2
bert-score==0.3.13
rouge-score==0.1.2
sentencepiece==0.1.99
langdetect==1.0.9
torch==2.0.1
huggingface-hub==0.16.4

# Database
pymongo==4.4.1
cassandra-driver==3.28.0

# Web and UI
streamlit==1.24.1
ipywidgets==8.0.6
pyngrok==6.0.0

# Environment and utilities
python-dotenv==1.0.0
requests==2.31.0